# Speed up output discovery