--- a/lib/galaxy/dependencies/pipfiles/default/pinned-dev-requirements.txt
+++ b/lib/galaxy/dependencies/pipfiles/default/pinned-dev-requirements.txt
@@ -37,6 +37,7 @@ pyparsing==2.4.7
 pytest-cov==2.10.0
 pytest-html==1.22.1
 pytest-metadata==1.8.0
+pytest-mock==3.1.1
 pytest-postgresql==2.3.0
 pytest-pythonpath==0.7.3
 pytest==5.4.3