--- a/lib/galaxy/model/store/discover.py
+++ b/lib/galaxy/model/store/discover.py
@@ -276,9 +276,14 @@ class ModelPersistenceContext(object):
         self.flush()
 
     def add_tags_to_datasets(self, datasets, tag_lists):
-        tag_session = self.tag_handler.create_tag_handler_session()
-        for dataset, tags in zip(datasets, tag_lists):
-            tag_session.add_tags_from_list(self.job.user, dataset, tags, flush=False)
+        if any(tag_lists):
+            # This works around SessionlessModelPersistenceContext not implementing a tag handler ...
+            # that's not better or worse than what we previously did in create_datasets
+            # TDOD: implement that or figure out why it is not implemented and find a better solution.
+            # Could it be that SessionlessModelPersistenceContext doesn't support tags?
+            tag_session = self.tag_handler.create_tag_handler_session()
+            for dataset, tags in zip(datasets, tag_lists):
+                tag_session.add_tags_from_list(self.job.user, dataset, tags, flush=False)
 
     @abc.abstractproperty
     def tag_handler(self):