This reduces the flushes from N datasets + 2 (f there are no tags to be crated, else it is 2*N datasets + 2 flushes)
to 3 flushes.

Creating 100 datasets via output discovery (splitting a text file with split_file_to_collection)
With this commit:
```
*** PROFILER RESULTS ***
discover_outputs (/Users/mvandenb/src/galaxy/lib/galaxy/jobs/__init__.py:1772)
function called 1 times

         792228 function calls (785314 primitive calls) in 1.009 seconds

   Ordered by: cumulative time, internal time, call count
   List reduced from 1266 to 40 due to restriction <40>

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
        1    0.000    0.000    1.013    1.013 __init__.py:1772(discover_outputs)
        1    0.000    0.000    1.013    1.013 __init__.py:1858(discover_outputs)
        1    0.000    0.000    1.009    1.009 output_collect.py:87(collect_dynamic_outputs)
        1    0.002    0.002    0.991    0.991 discover.py:192(populate_collection_elements)
        1    0.000    0.000    0.442    0.442 output_collect.py:266(add_datasets_to_history)
        1    0.000    0.000    0.441    0.441 __init__.py:1679(add_datasets)
      818    0.001    0.000    0.375    0.000 base.py:916(execute)
      818    0.001    0.000    0.373    0.000 elements.py:291(_execute_on_connection)
      818    0.004    0.000    0.373    0.000 base.py:1052(_execute_clauseelement)
7984/7675    0.006    0.000    0.334    0.000 attributes.py:698(get)
        3    0.001    0.000    0.330    0.110 session.py:2462(flush)
     9022    0.005    0.000    0.330    0.000 attributes.py:278(__get__)
        3    0.001    0.000    0.328    0.109 session.py:2515(_flush)
     2009    0.002    0.000    0.320    0.000 strategies.py:665(_load_for_state)
      100    0.002    0.000    0.316    0.003 discover.py:42(create_dataset)
      818    0.006    0.000    0.306    0.000 base.py:1159(_execute_context)
        3    0.000    0.000    0.292    0.097 unitofwork.py:402(execute)
        1    0.000    0.000    0.246    0.246 __init__.py:1688(<listcomp>)
      100    0.000    0.000    0.246    0.002 __init__.py:2611(get_total_size)
       19    0.000    0.000    0.224    0.012 persistence.py:184(save_obj)
      817    0.001    0.000    0.206    0.000 default.py:589(do_execute)
      817    0.201    0.000    0.206    0.000 {method 'execute' of 'psycopg2.extensions.cursor' objects}
      314    0.001    0.000    0.190    0.001 query.py:3425(_execute_and_instances)
        1    0.000    0.000    0.185    0.185 __init__.py:1700(__add_datasets_optimized)
       19    0.005    0.000    0.184    0.010 persistence.py:1039(_emit_insert_statements)
        1    0.000    0.000    0.181    0.181 mapping.py:2806(db_next_hid)
      4/3    0.000    0.000    0.179    0.060 session.py:893(begin)
      4/3    0.000    0.000    0.179    0.060 session.py:220(__init__)
      4/3    0.000    0.000    0.179    0.060 session.py:337(_take_snapshot)
      303    0.000    0.000    0.178    0.001 scoping.py:161(do)
      600    0.002    0.000    0.177    0.000 __init__.py:269(_invoke)
      207    0.000    0.000    0.175    0.001 state.py:640(_load_expired)
      107    0.001    0.000    0.173    0.002 loading.py:938(load_scalar_attributes)
      209    0.001    0.000    0.173    0.001 mapper.py:2797(_get_state_attr_by_column)
      207    0.000    0.000    0.170    0.001 strategies.py:753(_get_ident_for_use_get)
      107    0.000    0.000    0.170    0.002 loading.py:190(load_on_ident)
      107    0.001    0.000    0.169    0.002 loading.py:211(load_on_pk_identity)
      207    0.000    0.000    0.169    0.001 strategies.py:763(<listcomp>)
      107    0.000    0.000    0.168    0.002 query.py:3341(one)
      107    0.002    0.000    0.168    0.002 query.py:3306(one_or_none)
```

Without fix:
```
*** PROFILER RESULTS ***
discover_outputs (/Users/mvandenb/src/galaxy/lib/galaxy/jobs/__init__.py:1772)
function called 1 times

         1955870 function calls (1924658 primitive calls) in 2.463 seconds

   Ordered by: cumulative time, internal time, call count
   List reduced from 1289 to 40 due to restriction <40>

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
        1    0.000    0.000    2.472    2.472 __init__.py:1772(discover_outputs)
        1    0.000    0.000    2.472    2.472 __init__.py:1858(discover_outputs)
        1    0.000    0.000    2.467    2.467 output_collect.py:87(collect_dynamic_outputs)
        1    0.002    0.002    2.445    2.445 discover.py:190(populate_collection_elements)
      100    0.002    0.000    1.682    0.017 discover.py:42(create_dataset)
11065/10557    0.009    0.000    1.151    0.000 attributes.py:698(get)
    10222    0.007    0.000    1.146    0.000 attributes.py:278(__get__)
     1616    0.003    0.000    1.020    0.001 base.py:916(execute)
     1616    0.002    0.000    1.016    0.001 elements.py:291(_execute_on_connection)
     1616    0.010    0.000    1.014    0.001 base.py:1052(_execute_clauseelement)
      403    0.001    0.000    0.928    0.002 scoping.py:161(do)
      103    0.004    0.000    0.905    0.009 session.py:2462(flush)
      102    0.000    0.000    0.901    0.009 output_collect.py:212(flush)
      103    0.003    0.000    0.900    0.009 session.py:2515(_flush)
      505    0.002    0.000    0.812    0.002 state.py:640(_load_expired)
      405    0.003    0.000    0.807    0.002 loading.py:938(load_scalar_attributes)
      405    0.001    0.000    0.789    0.002 loading.py:190(load_on_ident)
      405    0.003    0.000    0.788    0.002 loading.py:211(load_on_pk_identity)
      405    0.000    0.000    0.782    0.002 query.py:3341(one)
      405    0.008    0.000    0.781    0.002 query.py:3306(one_or_none)
      812    0.003    0.000    0.713    0.001 query.py:3425(_execute_and_instances)
     1616    0.014    0.000    0.709    0.000 base.py:1159(_execute_context)
      103    0.003    0.000    0.708    0.007 unitofwork.py:402(execute)
     2008    0.004    0.000    0.683    0.000 strategies.py:665(_load_for_state)
      405    0.001    0.000    0.643    0.002 query.py:3400(__iter__)
     1615    0.001    0.000    0.517    0.000 default.py:589(do_execute)
     1615    0.504    0.000    0.516    0.000 {method 'execute' of 'psycopg2.extensions.cursor' objects}
      715    0.007    0.000    0.435    0.001 persistence.py:184(save_obj)
      408    0.002    0.000    0.368    0.001 mapper.py:2797(_get_state_attr_by_column)
      306    0.001    0.000    0.360    0.001 strategies.py:753(_get_ident_for_use_get)
      306    0.000    0.000    0.359    0.001 strategies.py:763(<listcomp>)
      101    0.000    0.000    0.342    0.003 output_collect.py:295(job_id)
      407    0.001    0.000    0.316    0.001 <string>:1(<lambda>)
      407    0.005    0.000    0.315    0.001 strategies.py:772(_emit_lazyload)
      615    0.001    0.000    0.313    0.001 unitofwork.py:585(execute)
      418    0.001    0.000    0.287    0.001 elements.py:405(compile)
      418    0.001    0.000    0.286    0.001 elements.py:470(_compiler)
      418    0.002    0.000    0.285    0.001 compiler.py:527(__init__)
      418    0.001    0.000    0.283    0.001 compiler.py:274(__init__)
  426/418    0.000    0.000    0.282    0.001 compiler.py:349(process)
```

and that's having a larger impact with more datasets to discover (The
covid workflows split a 10000 line file for instance).

We can use this pattern in other places as well, but I prefer going
through this one by one and opening separate PRs with profiling results.
Ideally we should attempt to batch flushes for mapped-over job creation as well,
which I hope will increase our job throughput.