@bernt-matthias 2020-06-27
Maybe one could add a test tool using the `maxseconds` attribute of `<test>`.