@mvdbeek 2020-06-27
> Maybe one could add a test tool using the `maxseconds` attribute of `<test>`.

That makes some assumption on how fast certain test should be. What do you think about a unit test that asserts the number of flushes instead ?