@bernt-matthias 2020-06-28
Yes, anything that helps detecting possible  degradations in the future is welcome.

Given that hardware is changing a runtime  limit makes only little sense anyway.